refactor: use keboola/php-db-import snowflake connection
