--- src/Keboola/DbExtractor/Snowflake/Exception.php
@@ -1,7 +1,0 @@
-<?php
-namespace Keboola\DbExtractor\Snowflake;
-
-class Exception extends \Exception
-{
-
-}
