+++ tests/Keboola/Extractor/SnowflakeTest.php
@@ -8,17 +8,6 @@ use Symfony\Component\Yaml\Yaml;
 
 class SnowflakeTest extends AbstractSnowflakeTest
 {
-    public function setUp()
-    {
-        if (!defined('APP_NAME')) {
-            define('APP_NAME', 'ex-db-snowflake');
-        }
-
-        $config = $this->getConfig();
-
-        $this->connection = new Connection($config['parameters']['db']);
-    }
-
     private function getUserDefaultWarehouse($user)
     {
         $sql = sprintf(
